# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SIGAT API
  version: 1.0.0
  description: Contrato inicial da API do SIGAT, alinhado aos fluxos consumidos pelo frontend.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: API versionada atrás do Nginx
...
paths:
...
  /missions:
    get:
      tags: [Operations]
      summary: Lista missões técnicas
      operationId: listMissions
      description: DITEL vê todas as missões; unidade vê apenas as atribuídas à sua unidade.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: unitId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [assigned, in_progress, completed, cancelled] }
      responses:
        '200':
          description: Lista de missões
          content: { application/json: { schema: { $ref: '#/components/schemas/MissionListResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Operations]
      summary: Cria missão técnica
      operationId: createMission
      description: Exclusivo para administrador DITEL.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MissionCreateRequest' }
      responses:
        '201':
          description: Missão criada
          content: { application/json: { schema: { $ref: '#/components/schemas/Mission' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /missions/{missionId}:
    patch:
      tags: [Operations]
      summary: Atualiza missão técnica
      operationId: updateMission
      description: Unidade pode iniciar e concluir; DITEL pode cancelar.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: missionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MissionUpdateRequest' }
      responses:
        '200':
          description: Missão atualizada
          content: { application/json: { schema: { $ref: '#/components/schemas/Mission' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /movements:
    get:
      summary: Listar movimentações autorizadas
      operationId: listMovements
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Movement list, content: { application/json: { schema: { $ref: '#/components/schemas/MovementListResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Solicitar transferência patrimonial
      operationId: createMovement
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateMovementRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Movement' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/AuditUnavailable' }
  /movements/{id}/decision:
    patch:
      summary: Decidir transferência patrimonial
      operationId: decideMovement
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/MovementId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MovementDecisionRequest' } } }
      responses:
        '200': { description: Updated movement, content: { application/json: { schema: { $ref: '#/components/schemas/Movement' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/AuditUnavailable' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sigat_session
  parameters:
    MovementId: { name: id, in: path, required: true, schema: { type: string } }
...
  responses:
    BadRequest:
      description: Dados inválidos
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
...
    Unauthorized:
      description: Sessão ausente ou inválida
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Forbidden:
      description: Usuário autenticado sem uma unidade válida para acessar o dashboard da unidade
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            invalidUnitScope:
              value:
                code: FORBIDDEN
                message: Escopo de unidade obrigatório.
    NotFound:
      description: Recurso não encontrado no escopo autorizado
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Conflict:
      description: Conflito de unicidade ou concorrência
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    AuditUnavailable:
      description: A mutação foi revertida porque a auditoria obrigatória não pôde ser registrada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            auditUnavailable:
              value:
                code: AUDIT_UNAVAILABLE
                message: Não foi possível registrar auditoria.
  schemas:
...
    Mission:
      type: object
      required: [id, title, description, type, status, unit, assignedBy, assignedTo, priority, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        type: { type: string, enum: [maintenance, installation, inspection, training, other] }
        status: { type: string, enum: [assigned, in_progress, completed, cancelled] }
        unit:
          type: object
          required: [id, name]
          properties:
            id: { type: string }
            name: { type: string }
            acronym: { type: string }
        equipment:
          type: array
          items:
            type: object
            required: [id, patrimony, type, model, brand]
            properties:
              id: { type: string }
              patrimony: { type: string }
              type: { type: string }
              model: { type: string }
              brand: { type: string }
        assignedBy:
          type: object
          required: [id, name, registration]
          properties:
            id: { type: string }
            name: { type: string }
            registration: { type: string }
        assignedTo:
          type: object
          required: [id, name, registration]
          properties:
            id: { type: string }
            name: { type: string }
            registration: { type: string }
        priority: { type: string, enum: [low, medium, high, critical] }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        notes:
          type: array
          items:
            type: object
            properties:
              author: { type: string }
              text: { type: string }
              createdAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MissionListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Mission' }
    MissionCreateRequest:
      type: object
      required: [title, description, unit]
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string, enum: [maintenance, installation, inspection, training, other] }
        priority: { type: string, enum: [low, medium, high, critical] }
        unit:
          type: object
          required: [id, name]
          properties:
            id: { type: string }
            name: { type: string }
            acronym: { type: string }
        equipment:
          type: array
          items:
            type: object
            required: [id, patrimony, type, model, brand]
            properties:
              id: { type: string }
              patrimony: { type: string }
              type: { type: string }
              model: { type: string }
              brand: { type: string }
    MissionUpdateRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [assigned, in_progress, completed, cancelled] }
        note: { type: string }
...
    Movement:
      type: object
      required: [id, type, equipmentId, origin, destination, requestedBy, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        type: { type: string, enum: [Transferência definitiva] }
        equipmentId: { type: string }
        origin: { $ref: '#/components/schemas/UnitContext' }
        destination: { $ref: '#/components/schemas/UnitContext' }
        requestedBy: { type: string }
        status: { type: string, enum: [Pendente, Aprovada, Rejeitada] }
        decisionReason: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MovementListResponse:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Movement' } } }
    CreateMovementRequest:
      type: object
      required: [equipmentId, destination]
      properties: { equipmentId: { type: string }, destination: { $ref: '#/components/schemas/UnitContext' } }
    MovementDecisionRequest:
      type: object
      required: [status, updatedAt]
      properties:
        status: { type: string, enum: [Aprovada, Rejeitada] }
        updatedAt: { type: string, format: date-time }
        reason: { type: string }
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: EQUIPMENT_NOT_FOUND }
        message: { type: string }
        details: {}
...
    UnitContext:
      type: object
      required: [id, name, acronym]
      properties:
        id: { type: string }
        name: { type: string }
        acronym: { type: string }
...